requirements.txt.txt
requests==2.20.0
python-dotenv==0.10.1